Reorganizes remote client classes
